--- src/sockTrader/core/exchanges/decorators/backtest.ts
@@ -1,1 +1,0 @@
-export {selfGeneratingOrders as backtest} from "./selfGeneratingOrders";
