--- src/sockTrader/core/exchanges/decorators/paperTrade.ts
@@ -1,1 +1,0 @@
-export {selfGeneratingOrders as paperTrade} from "./selfGeneratingOrders";
