+++ src/sockTrader/core/exchanges/hitBTC.ts
@@ -4,18 +4,16 @@ import config from "../../../config";
 import WebSocket, {Data} from "../connection/webSocket";
 import logger from "../logger";
 import Orderbook from "../orderbook";
-import {ICandle} from "../types/ICandle";
 import {ICandleInterval} from "../types/ICandleInterval";
 import {IConnection} from "../types/IConnection";
 import {IOrderbookData} from "../types/IOrderbookData";
 import {IResponseAdapter} from "../types/IResponseAdapter";
-import {IOrder, OrderSide} from "../types/order";
+import {OrderReportingBehaviour} from "../types/OrderReportingBehaviour";
 import {Pair} from "../types/pair";
 import BaseExchange from "./baseExchange";
 import HitBTCCommand from "./commands/hitBTCCommand";
-import {paperTrade} from "./decorators/paperTrade";
 import HitBTCAdapter from "./hitBTCAdapter";
-import {generateOrderId} from "./utils/utils";
+import HitBTCReportingBehaviour from "./orderReporting/hitBTCReportingBehaviour";
 
 export const CandleInterval: Record<string, ICandleInterval> = {
     ONE_MINUTE: {code: "M1", cron: "00 */1 * * * *"},
@@ -34,7 +32,6 @@ export const CandleInterval: Record<string, ICandleInterval> = {
  * The HitBTC class represent the HitBTC exchange
  * @see https://hitbtc.com/
  */
-@paperTrade
 export default class HitBTC extends BaseExchange {
     readonly adapter: IResponseAdapter = new HitBTCAdapter(this);
 
@@ -42,42 +39,8 @@ export default class HitBTC extends BaseExchange {
         return new WebSocket("wss://api.hitbtc.com/api/2/ws", 40 * 1000);
     }
 
-    adjustOrder(order: IOrder, price: number, qty: number): void {
-        if (this.orderManager.canAdjustOrder(order, price, qty)) {
-            const newOrderId = generateOrderId(order.pair);
-
-            this.send(new HitBTCCommand("cancelReplaceOrder", {
-                clientOrderId: order.id,
-                price,
-                quantity: qty,
-                requestClientId: newOrderId,
-                strictValidate: true,
-            }));
-        }
-    }
-
-    cancelOrder(order: IOrder): void {
-        this.orderManager.setOrderUnconfirmed(order.id);
-        this.send(new HitBTCCommand("cancelOrder", {clientOrderId: order.id}));
-    }
-
-    createOrder(pair: Pair, price: number, qty: number, side: OrderSide): void {
-        const orderId = generateOrderId(pair);
-        this.orderManager.setOrderUnconfirmed(orderId);
-
-        logger.info(`PRODUCTION ${side.toUpperCase()} ORDER! PRICE: ${price} SIZE: ${qty}`);
-        this.send(new HitBTCCommand("newOrder", {
-            clientOrderId: orderId,
-            price,
-            quantity: qty,
-            side,
-            symbol: pair,
-            type: "limit",
-        }));
-    }
-
     loadCurrencies(): void {
-        this.send(new HitBTCCommand("getSymbols"));
+        this.connection.send(new HitBTCCommand("getSymbols"));
     }
 
     /**
@@ -89,22 +52,12 @@ export default class HitBTC extends BaseExchange {
         const nonce: string = nanoid(32);
         const signature: string = crypto.createHmac("sha256", privateKey).update(nonce).digest("hex");
 
-        this.send(HitBTCCommand.createRestorable("login", {
+        this.connection.send(HitBTCCommand.createRestorable("login", {
             algo: "HS256",
             nonce,
             pKey: publicKey,
             signature,
         }));
-    }
-
-    onSnapshotCandles(pair: Pair, data: ICandle[], interval: ICandleInterval) {
-        this.getCandleManager(pair, interval, candles => this.emit("core.updateCandles", candles))
-            .set(data);
-    }
-
-    onUpdateCandles(pair: Pair, data: ICandle[], interval: ICandleInterval) {
-        this.getCandleManager(pair, interval, candles => this.emit("core.updateCandles", candles))
-            .update(data);
     }
 
     onSnapshotOrderbook({pair, ask, bid, sequence}: IOrderbookData) {
@@ -127,23 +80,23 @@ export default class HitBTC extends BaseExchange {
             period: interval.code,
         });
 
-        return this.send(command);
+        return this.connection.send(command);
     }
 
     subscribeOrderbook = (pair: Pair): void => {
         const command = HitBTCCommand.createRestorable("subscribeOrderbook", {symbol: pair.join("")});
-        this.send(command);
+        this.connection.send(command);
     };
 
     subscribeReports = (): void => {
         const command = HitBTCCommand.createRestorable("subscribeReports");
-        this.send(command);
+        this.connection.send(command);
     };
 
     protected onConnect(): void {
         super.onConnect();
 
-        this.getConnection().on("message", (data: Data) => this.adapter.onReceive(data));
+        this.connection.on("message", (data: Data) => this.adapter.onReceive(data));
         this.loadCurrencies();
 
         const auth = config.exchanges.hitbtc;
@@ -151,5 +104,9 @@ export default class HitBTC extends BaseExchange {
             logger.info("Live credentials are used!");
             this.login(auth.publicKey, auth.secretKey);
         }
+    }
+
+    protected getOrderReportingBehaviour(): OrderReportingBehaviour {
+        return new HitBTCReportingBehaviour(this.orderManager, this.connection, this);
     }
 }
